Server: project manager's `getUriForProjectFile()` is now `getFileByLocalUri()` and applies to all projects
We ignore what we believe is the "current" project as we're moving away from that concept, and instead find the file on any of the existing projects. We also return the whole file info object, not just its uri. In other words, this now works more like `getFile()`, but takes a local dependency uri.