Server: project files are consolidated when possible
In the event someone opens a file, and that file is an "include" file of another file that is opened later, the project manager now properly erases the original "Project" file and consolidates all files and its dependencies in the new project. This avoids having include files dangling as un-assembled projects.