Server: project scans for included dependencies is smarter
When scanning for a file's dependencies, we resolve its full expected uri before adding it as a dependency. This allows us to remove and add dependencies as needed, and to clean the projects more efficiencly.

This also adds support for the INCDIR pseudo-op where it just adds new path locations for possible include file targets.